Remember scroll bar location when changing tabs
The tabs on the left hand side scroll vertically when the contents
is long, e.g. when there are many events.

When switching tabs, e.g. between Events, Courses, or Results,
the position of the scroll bar is reset. E.g. if they had scrolled down to
the bottom event, viewed it, then gone back to the events tab,
the scroll bar would be reset to the top of the events list.

With this change, the scroll position is remembered for each tab between
switching by saving the scroll position for each tab as an attribute of
the tab contents div.